requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0

fastapi==0.135.3
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dotenv>=1.0.0
PyYAML==6.0.3
regex==2026.4.4
rich==14.3.3
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
yt-dlp>=2024.0.0
youtube-transcript-api>=0.6.1
